Fail on jinja2 render exception (#22)
This PR causes jcb to raise an a generic exception if a generic
exception is raised by the jinja2 renderer. Currently it causes only a
warning to be printed to standard out.

I don't know how many times my testing has failed because the yaml I get
rendered has simply "null" in it. I would rather catch the failure at
the rendering stage rather than at the point of running a JEDI
application with a bad yaml.