Change publishing on PyPi to on-release (#29)
In https://github.com/NOAA-EMC/jcb/pull/28 , I set GitHub Actions to
publish to PyPi on-push-to-develop. The problem is that every time the
GitHub action builds a new package for jcb, it will append the version
number, set in `__init__.py`, to the end of the name of the file that is
created. If this file is published to PyPi and that file, with its
version number already exists, the publish action will fail.

Thus, we need to update the version number every time we publish to
PyPi. In that case, we might as well release a tag in this repo that
matches the version number.

This PR changes the PyPi-publishing Github Action to publish-on-release
rather than publish-on-push-to-develop. When this PR is merged, I will
create a tag and release, which will publish this repo to PyPi.

I set the version number to `0.1.0`.